+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Added support for `ProcessIntString` parameters accepting `uint` values as returned by WMI/CIM calls
 * Added `-ArgumentEscaping` to `Start-ProcessWith` and `Start-ProcessEx`
 * Added `Credential` attribute transformer to transform username to PSCredential prompt on `-Credential` parameters
+* Add support for cancelling `Start-ProcessEx` and `Start-ProcessWith` with `-Wait`
 
 ## v0.4.0 - 2024-11-07